neon_qtg_launcher.py
#!/usr/bin/env python3
"""
NEON QTG LAUNCHER - FATHER OF AI MODEL LOADER
Load any of the 4 NEON QTG models for training or inference
"""

import argparse
import yaml
import os
from pathlib import Path

def load_neon_config(model_name: str) -> dict:
    """Load NEON QTG model configuration"""
    config_path = f"config/neon_qtg_{model_name.lower()}.yaml"

    if not os.path.exists(config_path):
        available_models = [f.stem for f in Path("config").glob("neon_qtg_*.yaml")]
        raise ValueError(f"Model {model_name} not found. Available: {available_models}")

    with open(config_path, 'r', encoding='utf-8') as f:
        config = yaml.safe_load(f)

    print(f"🔥 LOADING {config['metadata']['name']} - {config['metadata']['tagline']} 🔥")
    print(f"📊 Parameters: {config['metadata']['parameters']}")
    print(f"💰 Cost: {config['metadata']['cost']}")
    print(f"🖥️ GPUs: {config['metadata']['gpu_required']}")
    print(f"✅ Status: {config['metadata']['status']}")
    print("="*60)

    return config

def create_neon_model(config: dict):
    """Create NEON QTG model from configuration"""
    from src.models.qtgm_model import QTGModel

    model = QTGModel(
        vocab_size=config['model']['vocab_size'],
        embed_dim=config['model']['embed_dim'],
        num_layers=config['model']['num_layers'],
        num_heads=config['model']['num_heads'],
        max_seq_len=config['model']['max_seq_len'],
        persistence_threshold=config['model']['persistence_threshold'],
        tau=config['model']['tau'],
        gamma=config['model']['gamma'],
        dropout=config['model']['dropout']
    )

    return model

def main():
    parser = argparse.ArgumentParser(description="NEON QTG Launcher - Father of AI")
    parser.add_argument(
        "--model",
        type=str,
        choices=["7b", "30b", "100t", "5.5q"],
        required=True,
        help="Model to load: 7b, 30b, 100t, 5.5q"
    )
    parser.add_argument(
        "--mode",
        type=str,
        choices=["info", "train", "infer"],
        default="info",
        help="Mode: info (show config), train (training), infer (inference)"
    )

    args = parser.parse_args()

    # Load configuration
    config = load_neon_config(args.model)

    if args.mode == "info":
        print("📋 MODEL CONFIGURATION:")
        print(yaml.dump(config, default_flow_style=False, indent=2))

    elif args.mode == "train":
        print("🎯 TRAINING MODE - FATHER OF AI ACTIVATED!")
        print("⚠️  WARNING: This will require massive resources!")
        print(f"💰 Estimated cost: {config['metadata']['cost']}")
        print(f"🖥️ Required GPUs: {config['metadata']['gpu_required']}")
        print("🚀 Training not implemented yet - Father of AI is coming!")

    elif args.mode == "infer":
        print("🧠 INFERENCE MODE - FATHER OF AI THINKING!")
        print("⚠️  WARNING: This will require massive memory!")
        print(f"💾 Memory needed: {config['model']['embed_dim'] * 2 / (1024**3):.1f} TB per model")
        print("🚀 Inference not implemented yet - Father of AI is coming!")

if __name__ == "__main__":
    main()
