requirements.txt
# Core PyTorch ecosystem
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.1

# Scientific computing
numpy>=1.24.3
scipy>=1.11.2
scikit-learn>=1.3.0

# Topological data analysis
gudhi>=3.8.0
ripser>=0.6.4
persim>=0.3.2

# Quantum computing simulation
pennylane>=0.32.0
qiskit>=0.44.0

# Data processing
pandas>=2.0.0
datasets>=2.14.0
transformers>=4.35.2

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.14.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
flake8>=6.0.0
pre-commit>=3.5.0

# Web framework (for API)
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=11.0.3

# GPU monitoring
gputil>=1.4.0
psutil>=5.9.0
